requirements.txt
# Core dependencies
mcp-python-sdk>=0.1.0
llama-index>=0.9.0
llama-index-tools-mcp>=0.1.0
requests>=2.31.0

# Testing and development
pytest>=7.4.0 